Display all channels even though they are offline (a base for restarting channel server without restarting login)
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@919 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/LoginServer/Worlds.h
@@ -38,7 +38,7 @@ struct Channel {
 struct World {
 	string name;
 	hash_map <int, Channel *> channels;
-	int maxChannels;
+	size_t maxChannels;
 	char id;
 	char ribbon;
 	bool connected;
